Write Turtle File to Disk
Give up on using HTTP Clients to write the Dataset, because it fails to
load the amount of Triples in a reasonable time. The plan is to now
write a TTL File to disk and load it with the tdbloader2 into the TDB2
database. Also added a Script for spinning up the Fuseki Server.